dev: make context mgmt better
