+++ Cargo.toml
@@ -3,7 +3,7 @@ members = [".", "bevy_repl_derive"]
 
 [package]
 name = "bevy_repl"
-version = "0.3.1"
+version = "0.4.0"
 edition = "2024"
 description = "Add a REPL to headless Bevy applications"
 license = "MIT"
@@ -33,19 +33,21 @@ default = [
     "default_commands",
 ]
 dev = ["bevy/dynamic_linking"]
+derive = ["bevy_repl_derive", "clap/derive"]
+stdout = []
 default_commands = ["quit"]
 quit = []
 clear = []
 help = []
-derive = ["bevy_repl_derive", "clap/derive"]
 
 [[example]]
 name = "demo"
 path = "examples/demo.rs"
 
 [[example]]
-name = "minimal"
-path = "examples/minimal.rs"
+name = "stdout"
+path = "examples/stdout.rs"
+required-features = ["stdout"]
 
 [[example]]
 name = "fixed_update"
